+++ src/main/java/rocks/friedrich/engine_omega/GameLogic.java
@@ -46,7 +46,7 @@ public final class GameLogic {
     private static final int GRID_SIZE_METER_LIMIT = 100000;
     private static final int DEBUG_TEXT_SIZE = 12;
 
-    private static final float DESIRED_FRAME_DURATION = 0.016f;
+    private static final double DESIRED_FRAME_DURATION = 0.016f;
 
     private static final int NANOSECONDS_PER_SECOND = 1000000000;
 
@@ -62,7 +62,7 @@ public final class GameLogic {
      */
     private final Queue<Runnable> dispatchableQueue = new ConcurrentLinkedQueue<>();
 
-    private float frameDuration;
+    private double frameDuration;
 
     public GameLogic(RenderTarget render, Supplier<Scene> currentScene, Supplier<Boolean> isDebug) {
         this.render = render;
@@ -84,7 +84,7 @@ public final class GameLogic {
             Scene scene = this.currentScene.get();
 
             try {
-                float deltaSeconds = Math.min(2 * DESIRED_FRAME_DURATION, frameDuration);
+                double deltaSeconds = Math.min(2 * DESIRED_FRAME_DURATION, frameDuration);
 
                 scene.step(deltaSeconds, threadPoolExecutor::submit);
                 scene.getCamera().onFrameUpdate();
@@ -99,7 +99,7 @@ public final class GameLogic {
                 render();
 
                 frameEnd = System.nanoTime();
-                float duration = (float) (frameEnd - frameStart) / NANOSECONDS_PER_SECOND;
+                double duration = (double) (frameEnd - frameStart) / NANOSECONDS_PER_SECOND;
 
                 if (duration < DESIRED_FRAME_DURATION) {
                     try {
@@ -112,7 +112,7 @@ public final class GameLogic {
                 }
 
                 frameEnd = System.nanoTime();
-                frameDuration = ((float) (frameEnd - frameStart) / NANOSECONDS_PER_SECOND);
+                frameDuration = ((double) (frameEnd - frameStart) / NANOSECONDS_PER_SECOND);
 
                 frameStart = frameEnd;
             } catch (InterruptedException e) {
@@ -184,19 +184,19 @@ public final class GameLogic {
 
         Camera camera = scene.getCamera();
         Vector position = camera.getPosition();
-        float rotation = -camera.getRotation();
+        double rotation = -camera.getRotation();
 
         g.setClip(0, 0, width, height);
         g.translate(width / 2, height / 2);
 
-        float pixelPerMeter = camera.getZoom();
+        double pixelPerMeter = camera.getZoom();
 
         g.rotate(Math.toRadians(rotation), 0, 0);
         g.translate(-position.getX() * pixelPerMeter, position.getY() * pixelPerMeter);
 
-        int gridSizeInMeters = Math.round(GRID_SIZE_IN_PIXELS / pixelPerMeter);
-        float gridSizeInPixels = gridSizeInMeters * pixelPerMeter;
-        float gridSizeFactor = gridSizeInPixels / gridSizeInMeters;
+        int gridSizeInMeters = (int) Math.round(GRID_SIZE_IN_PIXELS / pixelPerMeter);
+        double gridSizeInPixels = gridSizeInMeters * pixelPerMeter;
+        double gridSizeFactor = gridSizeInPixels / gridSizeInMeters;
 
         if (gridSizeInMeters > 0 && gridSizeInMeters < GRID_SIZE_METER_LIMIT) {
             int windowSizeInPixels = Math.max(width, height);
@@ -225,7 +225,7 @@ public final class GameLogic {
 
             for (int x = startX; x <= stopX; x += gridSizeInMeters) {
                 for (int y = startY; y <= stopY; y += gridSizeInMeters) {
-                    g.drawString(x + " / " + -y, x * gridSizeFactor + 5, y * gridSizeFactor - 5);
+                    g.drawString(x + " / " + -y, (int) (x * gridSizeFactor + 5), (int) (y * gridSizeFactor - 5));
                 }
             }
         }
@@ -240,7 +240,7 @@ public final class GameLogic {
      */
     @Internal
     private void renderInfo(Graphics2D g, DebugInfo debugInfo) {
-        float frameDuration = debugInfo.getFrameDuration();
+        double frameDuration = debugInfo.getFrameDuration();
         int bodyCount = debugInfo.getBodyCount();
 
         Font displayFont = new Font("Monospaced", Font.PLAIN, DEBUG_TEXT_SIZE);
