+++ src/main/java/rocks/friedrich/engine_omega/actor/Animation.java
@@ -53,10 +53,10 @@ public class Animation extends Actor implements FrameUpdateListener {
 
     private final AnimationFrame[] frames;
 
-    private final float width;
-    private final float height;
+    private final double width;
+    private final double height;
 
-    private transient float currentTime;
+    private transient double currentTime;
     private transient int currentIndex;
 
     /**
@@ -64,7 +64,7 @@ public class Animation extends Actor implements FrameUpdateListener {
      */
     private final EventListeners<Runnable> onCompleteListeners = new EventListeners<>();
 
-    private Animation(AnimationFrame[] frames, float width, float height) {
+    private Animation(AnimationFrame[] frames, double width, double height) {
         super(() -> {
             if (frames.length < 1) {
                 throw new RuntimeException("Eine Animation kann nicht mit einem leeren Frames-Array initialisiert werden.");
@@ -115,7 +115,7 @@ public class Animation extends Actor implements FrameUpdateListener {
      * @see #getHeight()
      */
     @API
-    public float getWidth() {
+    public double getWidth() {
         return this.width;
     }
 
@@ -126,7 +126,7 @@ public class Animation extends Actor implements FrameUpdateListener {
      * @see #getWidth()
      */
     @API
-    public float getHeight() {
+    public double getHeight() {
         return this.height;
     }
 
@@ -144,7 +144,7 @@ public class Animation extends Actor implements FrameUpdateListener {
 
     @Internal
     @Override
-    public void onFrameUpdate(float deltaSeconds) {
+    public void onFrameUpdate(double deltaSeconds) {
         this.currentTime += deltaSeconds;
 
         AnimationFrame currentFrame = this.frames[currentIndex];
@@ -161,12 +161,12 @@ public class Animation extends Actor implements FrameUpdateListener {
     }
 
     @Override
-    public void render(Graphics2D g, float pixelPerMeter) {
+    public void render(Graphics2D g, double pixelPerMeter) {
         this.frames[currentIndex].render(g, width * pixelPerMeter, height * pixelPerMeter, false, false);
     }
 
     @API
-    public static Animation createFromSpritesheet(float frameDuration, String filepath, int x, int y, float width, float height) {
+    public static Animation createFromSpritesheet(double frameDuration, String filepath, int x, int y, double width, double height) {
         if (frameDuration <= 0) {
             throw new RuntimeException("Frame-Länge muss größer als 0 sein");
         }
@@ -196,7 +196,7 @@ public class Animation extends Actor implements FrameUpdateListener {
     }
 
     @API
-    public static Animation createFromImages(float frameDuration, float width, float height, String... filepaths) {
+    public static Animation createFromImages(double frameDuration, double width, double height, String... filepaths) {
         if (frameDuration <= 0) {
             throw new RuntimeException("Frame-Länge muss größer als 1 sein.");
         }
@@ -220,7 +220,7 @@ public class Animation extends Actor implements FrameUpdateListener {
      * @return Eine Animation aus allen Dateien, die mit dem Pfadpräfix beginnen.
      */
     @API
-    public static Animation createFromImagesPrefix(float frameDuration, float width, float height, String directoryPath, String prefix) {
+    public static Animation createFromImagesPrefix(double frameDuration, double width, double height, String directoryPath, String prefix) {
         // Liste mit den Pfaden aller qualifizierten Dateien
         ArrayList<String> allPaths = new ArrayList<>();
 
@@ -253,7 +253,7 @@ public class Animation extends Actor implements FrameUpdateListener {
     }
 
     @API
-    public static Animation createFromAnimatedGif(String filepath, float width, float height) {
+    public static Animation createFromAnimatedGif(String filepath, double width, double height) {
         GifDecoder gifDecoder = new GifDecoder();
         gifDecoder.read(filepath);
